+++ src/main/java/org/example/Solution/day02/model/CubeUpperLimit.java
@@ -1,7 +1,7 @@
 package org.example.Solution.day02.model;
 
-public record CubeUpperLimit(int blue, int green, int red) {
-    public int getPower() {
+public record CubeUpperLimit(long blue, long green, long red) {
+    public long getPower() {
         return blue * green * red;
     }
 }
